pyproject.toml
[project]
name = "evoo-agent"
version = "1.0.0"
description = "EVOO - Evolutionary Operations Optimizer: Autonomous SRE agent with learning capabilities"
requires-python = ">=3.12"
dependencies = [
    "agentex-sdk>=0.6.5",
    "temporalio>=1.5.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "termcolor>=2.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["project"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
